Updates dependencies and footer links.
Updates various dependencies, including `@tailwindcss/postcss`, `autoprefixer`, `lucide-react`, `react-router-dom`, `react`, `@types/react`, `@types/react-dom`, `html-webpack-plugin`, and `webpack`.

Changes a link in the footer to point to the CryptoCard page and reorders the footer links.

Updates a link on the ORIC solution page to point to the /joinus page.

Changes the text that is typed in the hero component.